Attach `InputControlled` to keyboard button labels.
This is instead of `SharedInputControlled`, so that each key can be
controlled per controller.

Issue #180

+++ crate/game_input_model/src/controller_id.rs
@@ -1,4 +1,4 @@
 /// Type for Controller ID.
 ///
 /// Better than stringly typed code.
-pub type ControllerId = u32;
+pub type ControllerId = usize;
